requirements.txt
Django>=5.2.8
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
pymongo==4.6.1



